+++ Package.swift
@@ -15,8 +15,8 @@ let package = Package(
     ],
     targets: [
         .binaryTarget(name: "FireworkVideo",
-                      url: "https://github.com/loopsocial/firework_ios_sdk/releases/download/v1.7.0/FireworkVideo-v1.7.0.xcframework.zip",
-                      checksum: "899af41a379cd59193d370a81b0e5f34392b2a34f17f828aa19c11248b8f07bf"),
+                      url: "https://github.com/loopsocial/firework_ios_sdk/releases/download/v1.8.0/FireworkVideo-v1.8.0.xcframework.zip",
+                      checksum: "aa6530274fb385648e0bbe486f38685f5be29527b23500d01de53d139581e80e"),
     ]
     
 )